+++ src/components/Nav/NavController.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import SideNav from './'
-import useSideNavReady from './useSideNavReady'
+import Nav from './'
+import useNavReady from './useNavReady'
 
 const NavWrapper = styled.div`
   position: relative;
@@ -11,14 +11,14 @@ const NavWrapper = styled.div`
 // Edit your links here
 const links = [
   { name: 'Under Armour', id: 'under-armour' },
-  { name: 'AMC', id: 'amc' },
+  { name: 'AMC Premiere', id: 'amc' },
   { name: 'Delta Dental', id: 'delta' },
 ]
 
 function NavController({ children }) {
   const CHILDREN_LENGTH = React.Children.count(children)
-  // Basically sideNav is "ready" when the first child is scrolled to the top of the screen
-  const sideNavReady = useSideNavReady(links[0].id)
+  // Basically Nav is "ready" when the first child is scrolled to the top of the screen
+  const NavReady = useNavReady(links[0].id)
 
   const [visibilityArray, setVisibilityArray] = React.useState(
     Array(CHILDREN_LENGTH).fill(0)
@@ -33,7 +33,7 @@ function NavController({ children }) {
   // Check to make sure that there is the same number of children that there are links
   if (links.length !== CHILDREN_LENGTH) {
     console.error(
-      'The number of links in the SideNav is not equal to the number of child components.'
+      'The number of links in the Nav is not equal to the number of child components.'
     )
     console.error('See the <NavConroller /> for further details.')
     return null
@@ -41,11 +41,7 @@ function NavController({ children }) {
 
   return (
     <NavWrapper>
-      <SideNav
-        isReady={sideNavReady}
-        links={links}
-        visibilityArray={visibilityArray}
-      />
+      <Nav isReady={NavReady} links={links} visibilityArray={visibilityArray} />
       {React.Children.map(children, (el, i) =>
         React.cloneElement(el, {
           id: links[i].id,
