+++ src/components/Projects/AMC/index.js
@@ -12,7 +12,7 @@ const StyledImage = styled(props => <Image {...props} />)`
 `
 
 const AMC = () => (
-  <div className="project begin-pres">
+  <div className="project pl4 pl5-l">
     <ImageQuery>
       {({ imageOne }) => [
         <Image
