+++ src/components/index.js
@@ -4,9 +4,9 @@ export { default as Footer } from './Footer'
 export { default as Hero } from './Hero'
 export { default as Image } from './Image'
 export { default as Layout } from './Layout'
-export { default as Nav } from './Nav'
+export { default as Header } from './Header'
 export { default as Nope } from './Nope'
 export { default as ScrollSection } from './ScrollSection'
 export { default as SEO } from './SEO'
-export { default as SideNav } from './SideNav'
+export { default as Nav } from './Nav'
 export { default as UnderArmour } from './Projects/UnderArmour'
