+++ src/styles/_global_styles.scss
@@ -14,7 +14,7 @@
   order: 10;
 }
 
-.lh-nav {
+.lh-header {
   line-height: 1.8;
 }
 
@@ -30,16 +30,16 @@
   font-family: $monospace;
 }
 
-.pr-nav {
+.pr-header {
   animation-delay: 500ms;
   animation-duration: 1s;
   animation-fill-mode: forwards;
-  animation-name: navIn;
+  animation-name: headerIn;
   animation-timing-function: $animation-function;
   transform: translateY(-150%);
 }
 
-@keyframes navIn {
+@keyframes headerIn {
   from {
     transform: translateY(-150%);
   }
@@ -49,7 +49,7 @@
   }
 }
 
-.nav-link {
+.header-link {
   position: relative;
 
   &::before {
@@ -140,10 +140,11 @@
 }
 
 .wrap-single {
-  margin: 0 auto;
+  margin: 0 auto 10vw;
   width: 71.5vw;
 
   @media (max-width: 767px) {
+    margin-bottom: 50px;
     width: 86%;
   }
 }
@@ -158,15 +159,6 @@
     font: 400 normal 20px/36px $sans-serif;
     padding: 50px 7vw 0;
     width: 100%;
-  }
-}
-
-.begin-pres {
-  margin: 0;
-  margin-top: 10vw;
-
-  @media (max-width: 767px) {
-    margin-top: 50px;
   }
 }
